refine: spec and lemma re-locate
