+++ ostd/libs/id-alloc/src/lib.rs
@@ -58,207 +58,6 @@ impl Inv for IdAlloc {
     }
 }
 
-/// `is_first_zero` is uniquely satisfied. Broadcast so it auto-fires whenever two
-/// `is_first_zero` facts arise.
-pub broadcast proof fn lemma_is_first_zero_unique(s: Seq<bool>, i: int, j: int)
-    requires
-        is_first_zero(s, i),
-        is_first_zero(s, j),
-    ensures
-        #![auto]
-        i == j,
-{
-    if i < j {
-        assert(false);
-    } else if j < i {
-        assert(false);
-    }
-}
-
-/// `first_zero_index(s)` itself satisfies `is_first_zero` (induction on `s.len()`).
-proof fn lemma_first_zero_index_is_first_zero(s: Seq<bool>)
-    ensures
-        is_first_zero(s, first_zero_index(s)),
-    decreases s.len(),
-{
-    if s.len() == 0 {
-    } else if !s[0] {
-    } else {
-        let sub = s.subrange(1, s.len() as int);
-        lemma_first_zero_index_is_first_zero(sub);
-        let i2 = first_zero_index(sub);
-        assert(is_first_zero(s, 1 + i2)) by {
-            assert forall|j: int| 0 <= j < 1 + i2 implies s[j] by {
-                if j == 0 {
-                } else {
-                    assert(s[j] == sub[j - 1]);
-                }
-            }
-            if 1 + i2 < s.len() {
-            }
-        }
-    }
-}
-
-/// If the prefix `[0, k)` of `s` is all `true`, then the first zero of `s` is
-/// `k` plus the first zero of the remainder (induction on `s.len()`).
-proof fn lemma_first_zero_index_after_true_prefix(s: Seq<bool>, k: int)
-    requires
-        0 <= k <= s.len(),
-        forall|j: int| #![trigger s[j]] 0 <= j < k ==> s[j],
-    ensures
-        first_zero_index(s) == k + first_zero_index(s.subrange(k, s.len() as int)),
-    decreases s.len(),
-{
-    if k == 0 {
-        assert(s.subrange(0, s.len() as int) =~= s) by {}
-    } else if s.len() == 0 {
-    } else {
-        let sub = s.subrange(1, s.len() as int);
-        lemma_first_zero_index_after_true_prefix(sub, k - 1);
-        assert(sub.subrange(k - 1, sub.len() as int) =~= s.subrange(k, s.len() as int)) by {}
-    }
-}
-
-/// Setting the bit at `k - 1` (the current first zero) to `true`, when the prefix
-/// `[0, k - 1)` is all `true`, advances the first zero to
-/// `k + first_zero_index(s.subrange(k, len))`.
-proof fn lemma_first_zero_index_advance_after_set(s: Seq<bool>, k: int)
-    requires
-        0 < k <= s.len(),
-        forall|j: int| 0 <= j < k - 1 ==> s[j],
-        k - 1 < s.len() ==> !s[k - 1],
-    ensures
-        first_zero_index(s.update(k - 1, true)) == k + first_zero_index(
-            s.subrange(k, s.len() as int),
-        ),
-{
-    let t = s.update(k - 1, true);
-    lemma_first_zero_index_after_true_prefix(t, k);
-    assert(t.subrange(k, s.len() as int) =~= s.subrange(k, s.len() as int)) by {}
-}
-
-/// Clearing a `true` bit at `i` moves the first zero to `min(first_zero_index(s), i)`.
-proof fn lemma_first_zero_index_clear(s: Seq<bool>, i: int)
-    requires
-        0 <= i < s.len(),
-        s[i],
-    ensures
-        first_zero_index(s.update(i, false)) == if first_zero_index(s) <= i {
-            first_zero_index(s)
-        } else {
-            i
-        },
-    decreases s.len(),
-{
-    let fz = first_zero_index(s);
-    lemma_first_zero_index_is_first_zero(s);
-    let t = s.update(i, false);
-    lemma_first_zero_index_is_first_zero(t);
-    if fz <= i {
-        assert(is_first_zero(t, fz)) by {
-            if fz < s.len() {
-                assert(!t[fz]) by {
-                    if fz == i {
-                        assert(t[fz] == false);
-                    } else {
-                        assert(t[fz] == s[fz]);
-                        assert(!s[fz]);
-                    }
-                }
-            }
-            assert(forall|j: int| 0 <= j < fz ==> t[j]) by {
-                assert(forall|j: int| 0 <= j < fz ==> s[j]);
-            }
-        }
-        lemma_is_first_zero_unique(t, first_zero_index(t), fz);
-    } else {
-        assert(is_first_zero(t, i)) by {
-            assert(!t[i]);
-            assert(forall|j: int| 0 <= j < i ==> t[j]) by {
-                assert(forall|j: int| 0 <= j < i ==> s[j]);
-            }
-        }
-        lemma_is_first_zero_unique(t, first_zero_index(t), i);
-    }
-}
-
-/// Clearing all bits in `[start, end)` moves the first zero to
-/// `min(first_zero_index(s), start)`.
-proof fn lemma_first_zero_index_clear_range(s: Seq<bool>, t: Seq<bool>, start: int, end: int)
-    requires
-        s.len() == t.len(),
-        0 <= start < end <= s.len(),
-        forall|j: int| #![trigger t[j]] 0 <= j < start ==> t[j] == s[j],
-        forall|j: int| #![trigger t[j]] start <= j < end ==> !t[j],
-        forall|j: int| #![trigger t[j]] end <= j < s.len() ==> t[j] == s[j],
-    ensures
-        first_zero_index(t) == if first_zero_index(s) <= start {
-            first_zero_index(s)
-        } else {
-            start
-        },
-{
-    let fz = first_zero_index(s);
-    lemma_first_zero_index_is_first_zero(s);
-    lemma_first_zero_index_is_first_zero(t);
-    if fz <= start {
-        assert(is_first_zero(t, fz)) by {
-            if fz < s.len() {
-                assert(!t[fz]) by {
-                    if fz < start {
-                        assert(t[fz] == s[fz]);
-                        assert(!s[fz]);
-                    }
-                }
-            }
-            assert(forall|j: int| 0 <= j < fz ==> t[j]) by {
-                assert(forall|j: int| 0 <= j < fz ==> s[j]);
-                assert(forall|j: int| 0 <= j < fz ==> t[j] == s[j]);
-            }
-        }
-        lemma_is_first_zero_unique(t, first_zero_index(t), fz);
-    } else {
-        assert(is_first_zero(t, start)) by {
-            assert(!t[start]);
-            assert(forall|j: int| 0 <= j < start ==> t[j]) by {
-                assert(forall|j: int| 0 <= j < start ==> s[j]);
-            }
-        }
-        lemma_is_first_zero_unique(t, first_zero_index(t), start);
-    }
-}
-
-/// Setting a `false` bit at `i` that is strictly past the first zero leaves the
-/// first zero unchanged.
-proof fn lemma_first_zero_index_set_after_first_zero(s: Seq<bool>, i: int)
-    requires
-        0 <= i < s.len(),
-        first_zero_index(s) < i,
-        !s[i],
-    ensures
-        first_zero_index(s.update(i, true)) == first_zero_index(s),
-{
-    let fz = first_zero_index(s);
-    lemma_first_zero_index_is_first_zero(s);
-    let t = s.update(i, true);
-    lemma_first_zero_index_is_first_zero(t);
-    assert(is_first_zero(t, fz)) by {
-        if fz < s.len() {
-            assert(!t[fz]) by {
-                assert(fz < i);
-                assert(t[fz] == s[fz]);
-                assert(!s[fz]);
-            }
-        }
-        assert forall|j: int| 0 <= j < fz implies t[j] by {
-            assert(t[j] == s[j]);
-            assert(s[j]);
-        }
-    }
-    lemma_is_first_zero_unique(t, first_zero_index(t), fz);
-}
-
 } // verus!
 #[verus_verify]
 impl IdAlloc {
