+++ verified_libs/vstd_extra/src/external/bitvec.rs
@@ -14,6 +14,7 @@
 //! The model is a `Seq<bool>`; the views below equate every executed `bitvec`
 //! operation to a `Seq` operation, so all reasoning in `id-alloc` stays at the
 //! `Seq<bool>` level.
+use crate::seq_extra::is_first_zero;
 use bitvec::{
     order::{BitOrder, Lsb0},
     slice::{BitSlice, BitSliceIndex},
@@ -294,27 +295,6 @@ pub broadcast axiom fn axiom_bitslice_get_range<'a, T: BitStore, O: BitOrder>(
             None => !(0 <= range.start <= range.end <= bitslice_view(bv).len()),
         },
 ;
-
-/// The index of the first `false` bit in `s`, or `s.len()` if every bit is `true`.
-pub open spec fn is_first_zero(s: Seq<bool>, i: int) -> bool {
-    &&& 0 <= i <= s.len()
-    &&& (forall|j: int| #![trigger s[j]] 0 <= j < i ==> s[j])
-    &&& (i < s.len() ==> !s[i])
-}
-
-/// Index of the first `false` bit, or `s.len()` if every bit is `true`. Defined
-/// recursively so it is deterministic and the SMT solver can unfold it.
-pub open spec fn first_zero_index(s: Seq<bool>) -> int
-    decreases s.len(),
-{
-    if s.len() == 0 {
-        0
-    } else if !s[0] {
-        0
-    } else {
-        1 + first_zero_index(s.subrange(1, s.len() as int))
-    }
-}
 
 /// The first index holding a `0` bit, counted from the start of the slice.
 pub assume_specification<T: BitStore, O: BitOrder>[ BitSlice::<T, O>::first_zero ](
